Simplified UserEntity class, moved WebSecurityConfig into \security package
